+++ src/Structure/Field/TimestampField.php
@@ -4,6 +4,7 @@ namespace Torr\PrismicApi\Structure\Field;
 
 use Symfony\Component\Validator\Constraints as Assert;
 use Torr\PrismicApi\Structure\Helper\FilterFieldsHelper;
+use Torr\PrismicApi\Transform\DataTransformer;
 use Torr\PrismicApi\Validation\DataValidator;
 
 /**
@@ -44,5 +45,13 @@ final class TimestampField extends InputField
 			),
 			$this->required ? new Assert\NotNull() : null,
 		]);
+	}
+
+	/**
+	 * @inheritDoc
+	 */
+	public function transformValue (mixed $data, DataTransformer $dataTransformer) : ?string
+	{
+		return $data;
 	}
 }